+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/util/Daita.kt
@@ -3,10 +3,10 @@ package net.mullvad.mullvadvpn.util
 import net.mullvad.mullvadvpn.compose.state.RelayListType
 
 fun shouldFilterByDaita(
-    isDaitaEnabled: Boolean,
+    daitaDirectOnly: Boolean,
     isMultihopEnabled: Boolean,
     relayListType: RelayListType,
 ) =
-    isDaitaEnabled &&
+    daitaDirectOnly &&
         (relayListType == RelayListType.ENTRY ||
             !isMultihopEnabled && relayListType == RelayListType.EXIT)