+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/viewmodel/location/SelectLocationViewModel.kt
@@ -18,9 +18,11 @@ import net.mullvad.mullvadvpn.compose.state.SelectLocationUiState
 import net.mullvad.mullvadvpn.lib.model.Constraint
 import net.mullvad.mullvadvpn.lib.model.CustomListId
 import net.mullvad.mullvadvpn.lib.model.RelayItem
+import net.mullvad.mullvadvpn.lib.model.Settings
 import net.mullvad.mullvadvpn.repository.CustomListsRepository
 import net.mullvad.mullvadvpn.repository.RelayListFilterRepository
 import net.mullvad.mullvadvpn.repository.RelayListRepository
+import net.mullvad.mullvadvpn.repository.SettingsRepository
 import net.mullvad.mullvadvpn.repository.WireguardConstraintsRepository
 import net.mullvad.mullvadvpn.usecase.FilterChipUseCase
 import net.mullvad.mullvadvpn.usecase.customlists.CustomListActionUseCase
@@ -34,6 +36,7 @@ class SelectLocationViewModel(
     private val relayListRepository: RelayListRepository,
     private val wireguardConstraintsRepository: WireguardConstraintsRepository,
     private val filterChipUseCase: FilterChipUseCase,
+    private val settingsRepository: SettingsRepository,
 ) : ViewModel() {
     private val _relayListType: MutableStateFlow<RelayListType> =
         MutableStateFlow(initialRelayListSelection())
@@ -44,30 +47,24 @@ class SelectLocationViewModel(
                 wireguardConstraintsRepository.wireguardConstraints,
                 _relayListType,
             ) { filterChips, wireguardConstraints, relayListSelection ->
-                SelectLocationUiState(
+                SelectLocationUiState.Data(
                     filterChips = filterChips,
                     multihopEnabled = wireguardConstraints?.isMultihopEnabled == true,
                     relayListType = relayListSelection,
                 )
             }
-            .stateIn(
-                viewModelScope,
-                SharingStarted.Lazily,
-                SelectLocationUiState(
-                    filterChips = emptyList(),
-                    multihopEnabled = false,
-                    relayListType = RelayListType.EXIT,
-                ),
-            )
+            .stateIn(viewModelScope, SharingStarted.Lazily, SelectLocationUiState.Loading)
 
     private val _uiSideEffect = Channel<SelectLocationSideEffect>()
     val uiSideEffect = _uiSideEffect.receiveAsFlow()
 
     private fun initialRelayListSelection() =
-        if (wireguardConstraintsRepository.wireguardConstraints.value?.isMultihopEnabled == true) {
-            RelayListType.ENTRY
-        } else {
-            RelayListType.EXIT
+        when {
+            settingsRepository.settingsUpdates.value?.daitaWithoutDirectOnly() == true ->
+                RelayListType.EXIT
+            wireguardConstraintsRepository.wireguardConstraints.value?.isMultihopEnabled == true ->
+                RelayListType.ENTRY
+            else -> RelayListType.EXIT
         }
 
     private fun filterChips() = _relayListType.flatMapLatest { filterChipUseCase(it) }
@@ -133,6 +130,10 @@ class SelectLocationViewModel(
     fun removeProviderFilter() {
         viewModelScope.launch { relayListFilterRepository.updateSelectedProviders(Constraint.Any) }
     }
+
+    private fun Settings.daitaWithoutDirectOnly() =
+        tunnelOptions.wireguard.daitaSettings.enabled &&
+            !tunnelOptions.wireguard.daitaSettings.directOnly
 }
 
 sealed interface SelectLocationSideEffect {